requirements.txt
streamlit
python-dotenv
agno
fastapi
pydantic
uvicorn
django
SpeechRecognition
gTTS
